♻️Refactor - 세미나 발표자 default 이미지로 이름 변경 #44
