+++ src/constants/seminar/seminarMemberData.ts
@@ -1,5 +1,5 @@
 import { SeminarMember } from "@/interfaces/seminar/seminarMember";
-import PresenterProfileImg from '@/images/seminar/presenter/presenter_profile_img.png';
+import PresenterProfileImg from '@/images/seminar/presenter/default_presenter.png';
 
 export const SEMINAR_MEMBER_DATA: SeminarMember = {
     id: 1,
